Update to new validator
