+++ README.md
@@ -15,7 +15,7 @@ Some notable features:
 - [LESS](http://lesscss.org) preconfigured out of the box to intelligently minify your external-facing CSS files. There's also built-in support for [Sass](https://sass-lang.com), and [Stylus](https://stylus-lang.com). Other CSS preprocessors can be used as well with a bit of extra configuration.
 - Built-in, easy to use interface for creating [Webpack](https://webpack.js.org/) bundles for modularizing and minifying your frontend JS.
 - Automatic server reloading when your backend code changes (via [nodemon](https://nodemon.io)) and automatic browser reloading when your frontend code changes (via [reload](https://github.com/alallier/reload)).
-- Automatic HTML validation in development mode of your post-server rendered HTML using a local instance of the [Nu HTML Checker](https://www.npmjs.com/package/vnu-jar). <img src='http://i.imgur.com/s4YUHNG.png' alt='' title='All life begins with Nu and ends with Nu...' width='16' height='16' style='image-rendering: -moz-crisp-edges; image-rendering: -o-crisp-edges; image-rendering: -webkit-optimize-contrast; image-rendering: crisp-edges; -ms-interpolation-mode: nearest-neighbor;'>
+- Automatic HTML validation in development mode of your post-server rendered HTML powered by [express-html-validator](https://github.com/rooseveltframework/express-html-validator).
 
 ![Teddy Roosevelt's facial hair is a curly brace.](https://github.com/rooseveltframework/generator-roosevelt/blob/master/generators/app/templates/statics/images/teddy.jpg "Teddy Roosevelt's facial hair is a curly brace.")
 
@@ -154,10 +154,6 @@ Roosevelt apps created with the app generator come with the following notable [n
   - Default shorthands:
     - `npm run pd`
   - Script is short for: `nodemon app.js --host-public`
-- `npm run kill-validator`: Finds the HTML validator process and kills it if it is running.
-  - Default shorthand:
-    - `npm run kv`
-  - Script is short for: `node ./node_modules/roosevelt/lib/scripts/killValidator.js`
 - `npm run clean`: Removes all build artifacts (symlinks and directories) auto-generated by Roosevelt. Will prompt to confirm before deleting any files.
   - Default shorthand:
     - `npm run c`
@@ -191,20 +187,6 @@ Roosevelt apps created with the app generator come with the following notable [n
   - Default shorthands:
     - `--raw`
     - `-r`
-- `node app.js --background-validator`: Forces the HTML validator to run as a detached background process.
-  - Default shorthand:
-    - `-b`
-- `node app.js --attach-validator`: Forces the HTML validator to run as an attached process.
-  - Default shorthand:
-    - `-a`
-- `node app.js --enable-validator-autokiller`: Forces the HTML validator autokiller to be enabled.
-  - Default shorthands:
-    - `--html-validator-autokiller`
-    - `-k`
-- `node app.js --disable-validator-autokiller`: Forces the HTML validator autokiller to be disabled.
-  - Default shorthands:
-    - `--no-autokiller`
-    - `-n`
 - `node app.js --host-public`: Forces Roosevelt to always host the [public folder](https://github.com/rooseveltframework/roosevelt#public-folder-parameters) even when `alwaysHostPublic` is set to false. Useful for testing production mode.
   - Default shorthands:
     - `--statics`
@@ -226,12 +208,6 @@ The following is a list of [environment variables](https://en.wikipedia.org/wiki
 - `NODE_PORT`: Default HTTP port to run your app on.
 - `HTTP_PORT`: Default HTTP port to run your app on. Takes precedence over `NODE_PORT`.
 - `HTTPS_PORT`: Default HTTPS port to run your app on.
-- `ROOSEVELT_VALIDATOR`:
-  - Set to `detached` to force the HTML validator to run as a detached background process.
-  - Set to `attached` to force the HTML validator to run as an attached process.
-- `ROOSEVELT_AUTOKILLER`:
-  - Set to `on` to spawn a process to kill the HTML validator if it is running in the background and idle for more than a certain amount of time. The timeout can be configured in [app behavior parameters](https://github.com/rooseveltframework/roosevelt#app-behavior-parameters).
-  - Set to `off`to disable the HTML validator autokiller.
 
 Environment variable precedence:
 
@@ -396,55 +372,40 @@ Resolves to:
 
   - Note: Automatically disabled during development mode.
 
-- `htmlValidator`: Parameters to send to [html-validator](https://github.com/zrrrzzt/html-validator#usage):
+- `htmlValidator`: Parameters to send to [express-html-validator](https://github.com/zrrrzzt/html-validator#usage):
 
   - `enable`: *[Boolean]* Enables or disables the built-in HTML validator.
 
     - Note: The validator is only available in development mode.
 
-  - `exceptions`: Sending a custom request header can disable the validator on a per request basis. The name of this request header and model value can be customized with this parameter.
+  - `exceptions`: A set of params that can be used to prevent validation in certain scenarios:
 
-    - Default: *[Object]*
+    - `header` *[String]*: A custom header that when set will disable the validator on a per request basis.
 
-        ```javascript
-        "exceptions": {
-          "requestHeader": "Partial",
-          "modelValue": "_disableValidator"
-        }
-        ```
+      - Default: `'Partial'`.
 
-  - `port`: *[Number]* Port to spawn the validator process on.
+    - `modelValue` *[String]*: An entry in your data model passed along with a `res.render` that when set will disable validation on the rendered HTML.
 
-  - `separateProcess`: *[Object]* How to run the validator:
+      - Default: `'_disableValidator'`
 
-    - `enable`: *[Boolean]* Run the validator as a detached background process.
+  - `validatorConfig` *[Object]*: [html-validate configuration](https://html-validate.org/usage/#configuration) that determines what errors the validator looks for.
 
-    - `autoKiller`: *[Boolean]* Spawns a process to kill the validator if it is running in the background and idle for more than a certain amount of time.
+    - Note: The full list of available validator rules can be found [here](https://html-validate.org/rules/).
 
-    - `autoKillerTimeout`: *[Number]* Time (in milliseconds) that the validator auto-killer process waits before it kills the validator running in the background.
-
-      - Note: You will see `"GET /roosevelt-dev-mode-ping HTTP/1.1" 404` in your HTTP logs when `autoKiller` is enabled if you allow your app to idle in development mode. This is a normal behavior that the autokiller uses to determine if it should kill the validator process.
-
-  - `showWarnings`: *[Boolean]* When set to true, shows HTML validation warnings in addition to errors.
+    - Note: This configuration can also be set by a `.htmlValidate.json` file placed in your app root directory.
 
   - Default: *[Object]*
 
-      ```json
-      {
-        "enable": true,
-        "separateProcess": {
-          "enable": true,
-          "autoKiller": true,
-          "autoKillerTimeout": 360000
-        },
-        "port": 48888,
-        "exceptions": {
-          "requestHeader": "Partial",
-          "modelValue": "_disableValidator"
-        },
-        "showWarnings": true
-      }
-      ```
+    ```json
+    {
+      "enable": true,
+      "exceptions": {
+        "requestHeader": "Partial",
+        "modelValue": "_disableValidator"
+      },
+      "validatorConfig": {}
+    }
+    ```
 
 - `formidable`: Settings to pass along to [formidable](https://github.com/felixge/node-formidable) using [formidable's API](https://github.com/felixge/node-formidable#api) for multipart form processing. Access files uploaded in your controllers by examining the `req.files` object. Roosevelt will remove any files uploaded to the `uploadDir` when the request ends automatically. To keep any, be sure to move them before the request ends. To disable multipart forms entirely, set this parameter to false.
