+++ lib/scripts/configAuditor.js
@@ -394,53 +394,13 @@ function configAudit (appDir) {
               case 'enable':
                 checkTypes(validatorParam[validatorKey], validatorKey, ['boolean'])
                 break
-              case 'separateProcess': {
-                checkTypes(validatorParam[validatorKey], validatorKey, ['object'])
-                const sepProcParam = validatorParam[validatorKey] || {}
-                for (const sepProcKey of Object.keys(sepProcParam)) {
-                  keyStack.push(sepProcKey)
-                  switch (sepProcKey) {
-                    case 'enable':
-                      checkTypes(sepProcParam[sepProcKey], sepProcKey, ['boolean'])
-                      break
-                    case 'autoKiller':
-                      checkTypes(sepProcParam[sepProcKey], sepProcKey, ['boolean'])
-                      break
-                    case 'autoKillerTimeout':
-                      checkTypes(sepProcParam[sepProcKey], sepProcKey, ['number'])
-                      break
-                    default:
-                      foundExtra(['rooseveltConfig', ...keyStack])
-                  }
-                  keyStack.pop(sepProcKey)
-                }
-                break
-              }
-              case 'port':
-                checkTypes(validatorParam[validatorKey], validatorKey, ['number', 'string'])
-                break
-              case 'showWarnings':
-                checkTypes(validatorParam[validatorKey], validatorKey, ['boolean'])
-                break
               case 'exceptions': {
                 checkTypes(validatorParam[validatorKey], validatorKey, ['object'])
-                const exceptionsParam = validatorParam[validatorKey] || {}
-                for (const exceptionKey of Object.keys(exceptionsParam)) {
-                  keyStack.push(exceptionKey)
-                  switch (exceptionKey) {
-                    case 'requestHeader':
-                      checkTypes(exceptionsParam[exceptionKey], exceptionKey, ['string'])
-                      break
-                    case 'modelValue':
-                      checkTypes(exceptionsParam[exceptionKey], exceptionKey, ['string'])
-                      break
-                    default:
-                      foundExtra(['rooseveltConfig', ...keyStack])
-                  }
-                  keyStack.pop(exceptionKey)
-                }
                 break
               }
+              case 'validatorConfig':
+                checkTypes(validatorParam[validatorKey], validatorKey, ['object'])
+                break
               default:
                 foundExtra(['rooseveltConfig', ...keyStack])
             }
@@ -569,12 +529,6 @@ function configAudit (appDir) {
             if (userScripts[script] !== undefined) {
               hasDevScript = true
             }
-            break
-          case 'kill-validator':
-            checkMissingScript(userScripts, script)
-            checkOutdatedScript(userScripts, script)
-            break
-          case 'kv':
             break
           case 'config-audit':
             checkMissingScript(userScripts, script)
