+++ lib/sourceParams.js
@@ -43,24 +43,6 @@ module.exports = (params, app) => {
     logging: false,
     // handle configuration edge cases
     transform: (params, flags) => {
-      // env variables first
-
-      // handle ROOSEVELT_VALIDATOR env var
-      if (process.env.ROOSEVELT_VALIDATOR === 'attached') {
-        params.htmlValidator.separateProcess.enable = false
-      } else if (process.env.ROOSEVELT_VALIDATOR === 'detached') {
-        params.htmlValidator.separateProcess.enable = true
-      }
-
-      // handle ROOSEVELT_AUTOKILLER env var
-      if (process.env.ROOSEVELT_AUTOKILLER === 'on') {
-        params.htmlValidator.separateProcess.autoKiller = true
-      } else if (process.env.ROOSEVELT_AUTOKILLER === 'off') {
-        params.htmlValidator.separateProcess.autoKiller = false
-      }
-
-      // cli flags next
-
       // handle --production-mode, --prod, -p cli flags
       if (flags.productionMode || flags.prod || flags.p) {
         params.mode = 'production'
@@ -77,22 +59,6 @@ module.exports = (params, app) => {
         params.htmlValidator.enable = true
       }
 
-      // handle --attach-validator, -a cli flags
-      if (flags.attachValidator || flags.a) {
-        params.htmlValidator.separateProcess.enable = false
-      } else if (flags.backgroundValidator || flags.b) {
-        // handle --background-validator, -b cli flags
-        params.htmlValidator.separateProcess.enable = true
-      }
-
-      // handle --disable-validator-autokiller, --no-autokiller, -n cli flags
-      if (flags.disableValidatorAutokiller || flags.autokiller === false || flags.n) {
-        params.htmlValidator.separateProcess.autoKiller = false
-      } else if (flags.enableValidatorAutokiller || flags.htmlValidatorAutokiller || flags.k) {
-        // handle --enable-validator-autokiller, --html-validator-autokiller, -k cli flags
-        params.htmlValidator.separateProcess.autoKiller = true
-      }
-
       // default mode param to production if its value is invalid
       if (params.mode !== 'production' && params.mode !== 'development') {
         params.mode = 'production'
@@ -103,7 +69,7 @@ module.exports = (params, app) => {
   }
 
   /**
-   * parameters are sourced via the following priority:
+   * Parameters are sourced via the following priority:
    * 1. command line arguments
    * 2. environment variables
    * 3. object passed to roosevelt function
@@ -140,23 +106,6 @@ module.exports = (params, app) => {
       enable: {
         default: defaults.htmlValidator.enable
       },
-      separateProcess: {
-        enable: {
-          default: defaults.htmlValidator.separateProcess.enable
-        },
-        autoKiller: {
-          default: defaults.htmlValidator.separateProcess.autoKiller
-        },
-        autoKillerTimeout: {
-          default: defaults.htmlValidator.separateProcess.autoKillerTimeout
-        }
-      },
-      port: {
-        default: defaults.htmlValidator.port
-      },
-      showWarnings: {
-        default: defaults.htmlValidator.showWarnings
-      },
       exceptions: {
         requestHeader: {
           default: defaults.htmlValidator.exceptions.requestHeader
@@ -164,6 +113,9 @@ module.exports = (params, app) => {
         modelValue: {
           default: defaults.htmlValidator.exceptions.modelValue
         }
+      },
+      validatorConfig: {
+        default: defaults.htmlValidator.validatorConfig
       }
     },
     formidable: {
