+++ package.json
@@ -25,34 +25,24 @@
   "dependencies": {
     "app-module-path": "~2.2.0",
     "body-parser": "~1.19.0",
-    "check-dependencies": "~1.1.0",
     "clean-css": "~4.2.3",
     "colors": "~1.4.0",
     "compression": "~1.7.4",
     "cookie-parser": "~1.4.5",
     "es6-template-strings": "~2.0.1",
-    "execa": "~4.0.1",
     "express": "~4.17.1",
-    "fkill": "~7.0.1",
     "formidable": "~1.2.2",
     "fs-extra": "~9.0.0",
     "html-minifier": "~4.0.0",
-    "html-validator": "~5.1.12",
     "klaw": "~3.0.0",
     "klaw-sync": "~6.0.0",
     "method-override": "~3.0.0",
     "morgan": "~1.10.0",
     "parent-require": "~1.0.0",
-    "pid-from-port": "~1.1.3",
-    "prismjs": "~1.20.0",
-    "ps-node": "~0.1.6",
-    "reload": "~3.0.4",
     "roosevelt-logger": "~0.2.2",
     "serve-favicon": "~2.5.0",
     "source-configs": "~0.3.3",
-    "tamper": "~1.1.0",
     "toobusy-js": "~0.5.1",
-    "vnu-jar": "~20.3.16",
     "webpack": "~4.43.0"
   },
   "devDependencies": {
@@ -71,6 +61,11 @@
     "stylus": "~0.54.7",
     "supertest": "~4.0.2",
     "teddy": "~0.4.28"
+  },
+  "optionalDependencies": {
+    "check-dependencies": "~1.1.0",
+    "express-html-validator": "~0.1.0",
+    "reload": "~3.0.4"
   },
   "repository": {
     "type": "git",
